requirements.txt
qiskit>=1.0.0
qiskit-aer>=0.14.0
qiskit-ibm-runtime>=0.20.0
amazon-braket-sdk>=1.82.0
amazon-braket-default-simulator>=1.23.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
matplotlib>=3.8.0
pyyaml>=6.0.1
networkx>=3.2.1
pylatexenc>=2.10
